requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
numpy==1.24.3
opencv-python==4.8.1.78
Pillow==10.1.0
pdf2image==1.16.3
pytesseract==0.3.10

# OCR Engines
paddlepaddle>=2.6.0
paddleocr>=2.7.0
easyocr==1.7.0

# Deep Learning & Computer Vision
torch==2.1.0
torchvision==0.16.0
scikit-image==0.21.0
scikit-learn==1.3.2

# Text Processing & Language Detection
langdetect==1.0.9
python-Levenshtein==0.21.1
fuzzywuzzy==0.18.0

# Image Processing
scipy==1.11.4
imutils==0.5.4

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pyyaml==6.0.1

